python/requirements.txt
protobuf==6.31.1
pyzmq==26.4.0
